V2 Diamond Mount Models
Thanks to Max Aalto for creating this back in August! 
Notes: Feedback so far is to mount the photodiode on the bottom, and widen some or the space for the light to pass through and create a shield over the photodiode to block external light. We hope to add this for V3 but if someone wants to jump in with designs, please feel free to!